+++ tests/test_plugin_setup.py
@@ -6,14 +6,14 @@ from kinto_pusher import __version__ as pusher_version
 
 class CapabilityTestView(BaseWebTest, unittest.TestCase):
 
-    def test_fxa_capability(self, additional_settings=None):
+    def test_pusher_capability(self):
         resp = self.app.get('/')
         capabilities = resp.json['capabilities']
         self.assertIn('pusher', capabilities)
         expected = {
             "version": pusher_version,
             "url": "https://github.com/Kinto/kinto-pusher",
-            "description": "Notify Pusher when somethings changes.",
+            "description": "Notify Pusher when something changes.",
             "app_id": "12345",
             "key": "demo-key"
         }